refactor: AutoFocusInputContainer 파일명 수정
